Working on player refactoring
