+++ src/components/call/renderer.tsx
@@ -1,0 +1,17 @@
+export const useCallRenderer = (
+  useSync: boolean,
+  line: LLCTCallLine,
+  time: number
+): {
+  activeLine?: boolean
+} => {
+  if (!useSync) {
+    return {}
+  }
+
+  const activeLine = line.start > 0 && line.start < time && line.end > time
+
+  return {
+    activeLine
+  }
+}
