+++ src/store/player/events.ts
@@ -1,0 +1,5 @@
+import { MusicPlayerState } from '@/@types/state'
+
+export interface PlayerRootEvents {
+  onStateChange: (state: MusicPlayerState) => void
+}
